saika/pyproject.toml
[project]
name = "saika"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = ["cryptography>=46.0.3", "pywebview>=6.1"]
